third_party/WebKit/LayoutTests/editing/inserting/insert-div-025.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<div id="log"></div>
<script>
test(() => assert_selection(
    [
        '<div contenteditable>',
            '<div>foo |bar</div>',
        '</div>',
    ].join(''),
    'insertParagraph',
    [
        '<div contenteditable>',
            '<div>foo\u{00A0}</div><div>|bar</div>',
        '</div>',
    ].join('')),
    'Keep a space after insert paragraph');
</script>
